+++ spotbugsTestCases/src/java/ghIssues/Issue4136.java
@@ -1,0 +1,10 @@
+package ghIssues;
+
+public class Issue4136 {
+
+    final String staticIsMissingPath;
+
+    {
+        staticIsMissingPath = "/search";
+    }
+}